refactor: Introduce DTOs to prevent entity exposure
Replace direct usage of Ticket entity in controller and service layers
with dedicated DTOs (TicketCreateRequestDTO, TicketResponseDTO,
TicketUpdateRequestDTO). This prevents exposing internal entity
structures to API consumers and improves separation of concerns.

Additionally:
- Change ticket price field from Double to BigDecimal for precision.
- Update .gitignore to exclude repomix output files.
- Remove test endpoint for expired reservation in controller.

Fixes #2

+++ src/main/java/com/omniticket/reservation_service/bootstrap/DataInitializer.java
@@ -1,5 +1,7 @@
 package com.omniticket.reservation_service.bootstrap;
 
+import java.math.BigDecimal;
+
 import com.omniticket.reservation_service.model.Ticket;
 import com.omniticket.reservation_service.model.TicketStatus;
 import com.omniticket.reservation_service.repository.TicketRepository;
@@ -26,7 +28,7 @@ public class DataInitializer implements ApplicationRunner {
             for (int i = 1; i <= 5; i++) {
                 Ticket ticket = new Ticket();
                 ticket.setSeatNumber("Sıra-A Koltuk-" + i);
-                ticket.setPrice(150.0 * i);
+                ticket.setPrice(BigDecimal.valueOf(150.0 * i));
                 ticket.setStatus(TicketStatus.AVAILABLE);
 
                 ticketRepository.save(ticket);
